Extract save loading out of GameState into GameStatePersistence
GameState.ts was 2,395 lines. Its loadState method — parse, ~47 legacy
field back-fills, the versioned migration chain, and the full default-state
literal — was ~370 of them and referenced only this.STORAGE_KEY, so it was
effectively a pure function trapped inside the class.

Moved it to GameStatePersistence.ts as loadPersistedState(storageKey), along
with the SAVE_VERSION / SAVE_MIGRATIONS / runSaveMigrations framework it uses.
GameState.ts re-exports SAVE_VERSION and runSaveMigrations so existing imports
(and tests/saveMigrations.test.ts) are unchanged; the persistence module
imports the GameState type back with `import type`, so there is no runtime
cycle. Behaviour is identical — a pure move, tsc- and test-verified.

GameState.ts 2,395 -> 2,025; the save-shape logic now lives in one focused,
441-line file that is the obvious place to look when a save fails to load.
